+++ hashutil/hashutil.go
@@ -1,4 +1,4 @@
-package util
+package hashutil
 
 import (
 	"crypto/sha1"
