+++ hashutil/hashutil_test.go
@@ -1,4 +1,4 @@
-package util
+package hashutil
 
 import (
 	"testing"
