+++ resource/resource.go
